% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the data from Rio de Janeiro}
\usage{
get_data(name)
}
\arguments{
\item{name}{Name of the transportation service}
}
\description{
Get the data from Rio de Janeiro
}
\examples{
data <- get_data(name = "supervia")
head(data)
}

